# api/admin.openapi.yaml
openapi: 3.1.0
info:
  title: Gatte management API
  version: 1.6.0
  summary: The backend every Gatte front talks to (design/adr/0040).
  description: |
    Served by `mcp-gateway admin` over UNIX sockets only. There is no TCP
    listener. HTTP/1.1, JSON bodies, snake_case fields, times in RFC 3339
    UTC.

    Two sockets, two processes, two directories, two disjoint route sets.
    Every operation below carries `x-gatte-socket`:

    - `operator`: `/run/mcp-gateway-admin/operator/operator.sock`, served
      by a process running as the gateway's service account; the file is
      group `gatte-operators`, `0660`, in a directory of its own.
    - `accounts`: `/run/mcp-gateway-admin/accounts/accounts.sock`, served
      only by a root process; the file is `root:root 0600`, or
      `root:ACCOUNT_GROUP 0660` when `[admin] account_group` delegates
      accounts (both settings are required; the backend refuses a mismatch
      at startup). Its directory is `root:root 0755`.

    A route of the other socket answers 404 `wrong_socket` with
    `details.socket`; a route that does not exist answers 404
    `unknown_route`.

    ## Who serves

    `pkg/adminapi` reads the server's credentials after connecting and
    refuses an accounts socket whose server is not uid 0 (under systemd the
    kernel reports systemd, which created the socket). `WithServerUID` pins
    the operator socket's server.

    ## Who the operator is

...
    `GET /v1/whoami` returns what the backend decided.

...

...
servers:
...
components:
  parameters:
    GatteFront:
      name: Gatte-Front
      in: header
      required: false
      description: Label of the calling front, for the operator row's tag. Display only.
      schema:
        type: string
        pattern: '^[a-z0-9._-]{1,32}$'
    Username:
      name: username
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Username'
  headers:
    NoStore:
      description: |
        Always `no-store` on a response that carries a one-time password.
        Advisory only on a UNIX socket: the protection a browser gets is the
        `no-store` that `pkg/frontkit` sends on every page.
      schema:
        type: string
        const: no-store
  responses:
    Error:
      description: |
        Any failure. `code` is stable and machine-readable; `message` is
        one English sentence for a person and may change.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: One English sentence for a person. Never carries a password.
        details:
          type: object
          description: |
            Code-specific facts. Known keys: `field` (invalid_argument;
            for maintenance `scope`, `upstream`, `message` or `until`),
            `upstream` (not_found on a maintenance request),
            `observed_fingerprint` (fingerprint_required, fingerprint_mismatch,
            fingerprint_moved), `status` (fingerprint_mismatch), `username`
            (account_exists, not_found, account_not_managed), `group`
            (group_not_mapped, account_not_managed), `socket`
            (wrong_socket: the socket that serves this route), `path` and
            `problem` (config_unavailable: the file or directory that failed
            the check, and `not_root_owned`, `group_or_other_writable`,
            `symlink`, `not_regular` or `does_not_load`). More keys may appear.
          additionalProperties: true
...
    Username:
      type: string
      pattern: '^[a-z0-9][a-z0-9._-]{0,63}$'
...
    OperatorRow:
      type: object
      description: The operator row this action wrote to the audit trail.
      required: [identity, tool, reason, timestamp]
      properties:
        identity:
          type: string
          examples: ['(operator:alice)']
        tool:
          type: string
          examples: ['(tool approve)']
        reason:
          type: string
        timestamp:
          type: string
          format: date-time
    ActionResult:
      type: object
      description: Common part of every state-changing response.
      required: [changed, recorded, messages]
      properties:
        changed:
          type: boolean
          description: false when the request was already true (already blocked, already pending); no row is written then.
        recorded:
          type: boolean
          description: |
            true when `audit` holds the operator row of this change. false
            with `changed` true means the change is in force and the trail
            could not record it (warning `audit_write_failed`). false with
            `changed` false means nothing needed recording.
        audit:
          $ref: '#/components/schemas/OperatorRow'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/Warning'
        messages:
          type: array
          description: |
            What the CLI would have printed, one line per item, for a
            "details" disclosure. English, not for localisation; never
            carries a password.
          items:
            type: string
...
    Group:
      type: object
      required: [name, role]
      properties:
        name: {type: string}
        role: {type: string}
    Role:
      type: object
      required: [name, groups, tools]
      properties:
        name: {type: string}
        groups:
          type: array
          items: {type: string}
        tools:
          type: array
          description: '`backend.tool` names and patterns this role grants, sorted.'
          items: {type: string}
    Seen:
      type: object
      required: [identity, last_call, calls, blocked]
      properties:
        identity: {type: string}
        name: {type: string}
        last_call: {type: string, format: date-time}
        calls: {type: integer}
        blocked: {type: boolean}
    People:
      type: object
      required: [roles, groups, seen]
      properties:
        roles:
          type: array
          items: {$ref: '#/components/schemas/Role'}
        groups:
          type: array
          description: Every `[group_to_role]` key, sorted.
          items: {$ref: '#/components/schemas/Group'}
        seen:
          type: array
          description: Analysts the trail has seen, most recent first. Rows in parentheses (gateway, operator, unauthenticated) are left out.
          items: {$ref: '#/components/schemas/Seen'}
    ConnectScript:
      type: object
      required: [os, name, filename, run, content]
      properties:
        os: {type: string, enum: [macos, linux, windows]}
        name: {type: string, examples: [macOS]}
        filename: {type: string, examples: [connect-gatte.sh]}
        run: {type: string, description: The command the analyst runs after downloading.}
        content: {type: string, description: The whole script. Carries no secret.}
    Connect:
      type: object
      required: [ready]
      properties:
        ready: {type: boolean}
        missing:
          type: array
          description: When not ready, the `[connect]` keys that are missing or invalid, e.g. `[client_id, callback_port]`.
          items: {type: string}
        url: {type: string, examples: ['https://gateway.example.internal/mcp']}
        client_id: {type: string}
        callback_port: {type: integer}
        server_name: {type: string, examples: [gatte]}
        username: {$ref: '#/components/schemas/Username'}
        has_ca: {type: boolean}
        command:
          type: string
          description: The one `claude mcp add` line for a machine that needs no CA.
        scripts:
          type: array
          items: {$ref: '#/components/schemas/ConnectScript'}
    Account:
      type: object
      description: An IdP account. The password hash is never returned.
      required: [username, display_name, groups, disabled]
      properties:
        username: {$ref: '#/components/schemas/Username'}
        display_name: {type: string}
        email: {type: string}
        groups:
          type: array
          items: {type: string}
        disabled: {type: boolean}
    AssignableGroup:
      type: object
      required: [name, role, tools]
      properties:
        name: {type: string}
        role: {type: string}
        tools:
          type: array
          items: {type: string}
    AccountDraft:
      type: object
      additionalProperties: false
      properties:
        display_name: {type: string}
        username:
          type: string
          description: Empty asks the backend to suggest one from display_name.
        email: {type: string}
        groups:
          type: array
          items: {type: string}
    AccountCheck:
      type: object
      description: The wizard's validation, with no side effect.
      required: [ok, username, suggested, problems, groups]
      properties:
        ok: {type: boolean}
        username: {type: string}
        suggested: {type: boolean, description: true when username was derived from display_name.}
        problems:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field: {type: string, enum: [display_name, username, email, groups]}
              code: {type: string, examples: [invalid_argument, account_exists, group_not_mapped]}
              message: {type: string}
        groups:
          type: array
          description: The requested groups that map to a role, with the role's tools.
          items: {$ref: '#/components/schemas/AssignableGroup'}
    NewAccount:
      type: object
      required: [username, display_name, groups]
      additionalProperties: false
      properties:
        username: {$ref: '#/components/schemas/Username'}
        display_name: {type: string, maxLength: 100}
        email: {type: string}
        groups:
          type: array
          description: Each must be a key of `[group_to_role]`.
          items: {type: string}
    GroupsRequest:
      type: object
      required: [groups]
      additionalProperties: false
      properties:
        groups:
          type: array
          items: {type: string}
    OffboardRequest:
      type: object
      additionalProperties: false
      properties:
        subject:
          type: string
          description: |
            The person's analyst identity, as the trail's `analyst_identity`
            (the token's `sub`). Optional: without it no gateway block is
            placed, and `remaining` says so. With it, a peer that is not root
            must also be in `[admin] operator_group` (`forbidden_peer`).
        reason:
          type: string
          description: Free text; recorded in each row after the front tag.
    OffboardResult:
      allOf:
        - $ref: '#/components/schemas/ActionResult'
        - type: object
          required: [account, blocked, disabled, rows, remaining]
          properties:
            account: {$ref: '#/components/schemas/Account'}
            subject: {type: string}
            blocked:
              type: boolean
              description: The subject is blocked now (placed by this call, or already).
            disabled:
              type: boolean
              description: The account is disabled now (by this call, or already).
            rows:
              type: array
              description: "The rows written, in order: `(access block)`, `(account disable)`, `(account offboard)`."
              items: {$ref: '#/components/schemas/OperatorRow'}
            remaining:
              type: array
              description: What the operator still does by hand, in order. Never empty; the first is revoking the sessions at the identity provider.
              items: {type: string}
    AccountResult:
      allOf:
        - $ref: '#/components/schemas/ActionResult'
        - type: object
          required: [account]
          properties:
            account: {$ref: '#/components/schemas/Account'}
    PasswordResult:
      allOf:
        - $ref: '#/components/schemas/ActionResult'
        - type: object
          required: [account, one_time_password]
          properties:
            account: {$ref: '#/components/schemas/Account'}
            one_time_password:
              type: string
              description: |
                20 characters, about 117 bits. Returned once, in this field of
                this response only, also when `recorded` is false; stored only
                as an argon2id hash; never in `messages`, `warnings`, an error,
                the trail or the backend's log. A front shows it once and keeps
                no copy.
...
paths:
...
  /v1/people:
    get:
      operationId: people
      tags: [people]
      x-gatte-socket: [operator]
      summary: Roles with their groups and tools, and who the trail has seen.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/People'}
        default: {$ref: '#/components/responses/Error'}
  /v1/connect:
    get:
      operationId: connect
      tags: [connect]
      x-gatte-socket: [operator]
      summary: How an analyst connects Claude Code, with the three scripts rendered (design/adr/0039).
      parameters:
        - {name: username, in: query, schema: {$ref: '#/components/schemas/Username'}, description: Written into the scripts' comments and final hint.}
      responses:
        '200':
          description: OK. `ready` false with `missing` when `[connect]` is incomplete.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Connect'}
        default: {$ref: '#/components/responses/Error'}
  /v1/connect/script:
    get:
      operationId: connectScript
      tags: [connect]
      x-gatte-socket: [operator]
      summary: One script as a file.
      parameters:
        - {name: os, in: query, required: true, schema: {type: string, enum: [macos, linux, windows]}}
        - {name: username, in: query, schema: {$ref: '#/components/schemas/Username'}}
      responses:
        '200':
          description: The script. `Content-Disposition` names the file.
          headers:
            Content-Disposition:
              schema: {type: string, examples: ['attachment; filename="connect-gatte.sh"']}
          content:
            text/plain:
              schema: {type: string}
        default:
          description: connect_not_configured, invalid_argument (unknown os), ...
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/groups:
    get:
      operationId: assignableGroups
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: The groups an account may be given (keys of `[group_to_role]`), with role and tools.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [groups]
                properties:
                  groups:
                    type: array
                    items: {$ref: '#/components/schemas/AssignableGroup'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Accounts in `[idp] users_file`.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [accounts]
                properties:
                  accounts:
                    type: array
                    items: {$ref: '#/components/schemas/Account'}
        default: {$ref: '#/components/responses/Error'}
    post:
      operationId: addAccount
      x-gatte-idempotent: false
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Create an account with a generated one-time password. Writes `(account add)`.
      parameters: [{$ref: '#/components/parameters/GatteFront'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/NewAccount'}
      responses:
        '201':
          description: Created (also when `recorded` is false). Warning `idp_reload` says the IdP applies it when it re-reads the file.
          headers:
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PasswordResult'}
        default:
          description: |
            invalid_argument, group_not_mapped, account_exists, ... NOT
            idempotent: after `internal` or a dropped connection, GET the
            account; if it exists, the password was lost with the answer, so
            call resetAccountPassword instead of retrying.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorResponse'}
  /v1/account-check:
    post:
      operationId: checkAccount
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Validate a draft and suggest a username. No side effect, nothing recorded.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AccountDraft'}
      responses:
        '200':
          description: OK, whether or not the draft is valid; read `ok` and `problems`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountCheck'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts/{username}:
    get:
      operationId: getAccount
      tags: [accounts]
      x-gatte-socket: [accounts]
      parameters: [{$ref: '#/components/parameters/Username'}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Account'}
        default: {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteAccount
      x-gatte-idempotent: false
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Remove the account from the identity provider's users file (1.4.0, `account_delete`). Writes `(account delete)`.
      description: |
        Same reach as disable (`account_not_managed`). A second call answers
        `not_found`. The trail keeps every row; no session is ended and no
        block placed or lifted.
      parameters:
        - {$ref: '#/components/parameters/Username'}
        - {$ref: '#/components/parameters/GatteFront'}
      responses:
        '200':
          description: Deleted; `account` is the account as it was.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts/{username}/offboard:
    post:
      operationId: offboardAccount
      x-gatte-idempotent: true
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Block the person in the gateway and disable their account, in one call (1.4.0, `offboard`).
      description: |
        The block first, through a child running as the database's owner (the
        root process still never opens the service account's database), then
        the disable. Writes `(access block)`, `(account disable)` and a summary
        `(account offboard)`. Nothing changes when the block cannot be placed;
        when the block is in force and the disable failed, 200 with warning
        `offboard_incomplete`. Same reach as disable (`account_not_managed`).
        Nothing at the identity provider's sessions is touched: `remaining`
        says what is left.
      parameters:
        - {$ref: '#/components/parameters/Username'}
        - {$ref: '#/components/parameters/GatteFront'}
      requestBody:
        required: false
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OffboardRequest'}
      responses:
        '200':
          description: Offboarded, or already (`changed` false).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OffboardResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts/{username}/groups:
    post:
      operationId: setAccountGroups
      x-gatte-idempotent: true
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Replace the account's groups. Writes `(account groups)`.
      parameters:
        - {$ref: '#/components/parameters/Username'}
        - {$ref: '#/components/parameters/GatteFront'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GroupsRequest'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts/{username}/disable:
    post:
      operationId: disableAccount
      x-gatte-idempotent: true
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Refuse new logins. Does not end a token already issued (use access block). Writes `(account disable)`.
      parameters:
        - {$ref: '#/components/parameters/Username'}
        - {$ref: '#/components/parameters/GatteFront'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts/{username}/enable:
    post:
      operationId: enableAccount
      x-gatte-idempotent: true
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Allow logins again. Writes `(account enable)`.
      parameters:
        - {$ref: '#/components/parameters/Username'}
        - {$ref: '#/components/parameters/GatteFront'}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccountResult'}
        default: {$ref: '#/components/responses/Error'}
  /v1/accounts/{username}/reset-password:
    post:
      operationId: resetAccountPassword
      x-gatte-idempotent: false
      tags: [accounts]
      x-gatte-socket: [accounts]
      summary: Replace the password with a generated one-time password. Writes `(account reset password)`.
      description: |
        Safe to repeat; each call invalidates the previous one-time password, so only the last answer's password works.
        Groups, disable, enable and reset-password reach only accounts Gatte manages (at least one group, every one
        in `[group_to_role]`) unless the peer is root; any other account is `account_not_managed`.
      parameters:
        - {$ref: '#/components/parameters/Username'}
        - {$ref: '#/components/parameters/GatteFront'}
      responses:
        '200':
          description: OK
          headers:
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PasswordResult'}
        default: {$ref: '#/components/responses/Error'}